// types.ts
export type ResponseType = {
  success: boolean;
  data?: any;
  msg?: string;
};
export type InventoryType = {
  id?: string;
  name: string;
  category: string;
  brand?: string;
  purchasedPrice: number;
  sellingPrice: number;
  expireDate: Date | Timestamp | string;
  quantity?: number;
  description?: string;
  image?: any | null;
  created?: string;
  updated?: string;
  ownerName?: string;
};